splitting tests into files according to which files they test. Refactoring tests to make more sense
