+++ src/case_gather/test_data/caselist.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<málaskrá>
+    <mál málsnúmer='1' þingnúmer='146' málsflokkur='A'>
+        <málsheiti>fjárlög 2017</málsheiti>
+    </mál>
+    <mál málsnúmer='2' >
+        <málsheiti>ýmsar forsendur fjárlagafrumvarps 2017</málsheiti>
+    </mál>
+    <mál málsnúmer='3'>
+        <málsheiti>sálfræðiþjónusta í framhaldsskólum</málsheiti>
+    </mál>
+</málaskrá>
